requirements.txt
cryptography==45.0.7
fastapi==0.116.1
langchain==0.3.27
langchain_groq==0.3.7
langgraph==0.6.6
numpy==2.3.2
psycopg==3.2.9
psycopg2_binary==2.9.10
psycopg_pool==3.2.6
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.0.0
setuptools==80.9.0
starlette==0.47.3
torch==2.7.1
